-Removed skipReturnCheck mechanism from calling convention.  It's no longer stored in a continuation, nor passed during an invocation.  Instead, an L2 call site is responsible for checking the returned value against the expected type if the VM was unable to prove its compliance at L1 -> L2 translation time. -L2 translations of method invocations now take into account the expanded paths of the method's LookupTree.  Unexpanded paths and paths that lead to an invalid lookup are now routed to the slower (non-inline), generic polymorphic lookup mechanism.  A CallSiteHelper carries much of the call state information, such as control flow merging labels. -When code generating L2 and starting a new block, if that block only has one predecessor, and that predecessor reaches it via an L2_JUMP, discard the block and jump and continue code generation in the predecessor block. -Strengthened JavaDoc @param and @return standard (still only warnings).

+++ src/com/avail/descriptor/FloatDescriptor.java
@@ -437,7 +437,7 @@ extends AbstractNumberDescriptor
 		final AvailObject object,
 		final @Nullable Class<?> ignoredClassHint)
 	{
-		return Float.valueOf(getFloat(object));
+		return getFloat(object);
 	}
 
 	@Override @AvailMethod
